There seems to be a memory leak. Usage increase steadily each time another test is run. Not sure why, I tracked the call to delete simulation and it seems to delete everything. I'm going to test on the workingOnFInalVersion branch quick to see if its just since the merge.

+++ Comp 524 Project/src/TriangleProblemGraph.cpp
@@ -83,7 +83,7 @@ inline void TriangleProblemGraph::runTestCase() {
 
 inline void TriangleProblemGraph::block2() {
 
-	int sum { A + B + C };
+	long long sum { A + B + C };
 
 	if( 2*A < sum && 2*B < sum && 2*C < sum ) {
 		testCase->addPredicateCoverage(predicates::B2_TTT);
